+++ src/main/scala/org/podval/tools/publish/markup/EntityReference.scala
@@ -1,7 +1,6 @@
 package org.podval.tools.publish.markup
 
-import org.podval.xml.XmlExtras
-import org.podval.xml.codec.XmlCodec
+import org.podval.xml.{XmlCodec, XmlExtras}
 import zio.blocks.schema.{Modifier, Schema}
 import zio.blocks.typeid.TypeId